aqui si ya esta la tabla pero no lista
