+++ src/App.js
@@ -3,12 +3,14 @@ import './App.css';
 import Menu from './components/MENU/Menu';
 import TablaPage from './pages/TablasPage/TablaPage';
 import Result from './pages/RESULT/Result';
+import Tabla from './components/TablaAlumno/Tabla';
 function App() {
   return (
     <div className="App">
       <Menu />
       {/* <TablaPage /> */}
-      <Result />
+      {/* <Result /> */}
+      <Tabla />
     </div>
   );
 }
